Jana/src-tauri/resources/jana-agents/requirements.txt
autogen-agentchat
autogen-ext[openai]
autogen-ext[docker]
python-dotenv
flask
hypercorn
asgiref
fastapi
uvicorn
paramiko
tavily-python